Reentrancy: introduce a temporary emu global
This patch introduces a global emu constant, equal in value to &M, and
switches all references to the M global to use emu instead. Subsequent
commits are going to gradually introduce `emu` as a parameter in each
individual function. The purpose of this is to reduce the number of
lines that have to change in the conversion process.

The following GNU sed commands were used to generate this patch:

    sed -i -e 's!M\.!emu->!g ; s!&M!emu!g' *.c
    sed -i -e 's!M\.!(emu)->!g ; s!&M!(emu)!g' include/*.h
    sed -i -e 's!emu = emu!emu = \&M!' api.c
    sed -i -e '/extern x86emu_t M;/ a static x86emu_t *const emu = &M;' include/x86emu_int.h
    sed -i -e '/if(emu)/ s,emu,emu_,g' decode.c
    sed -i -e '/unsigned x86emu_run/ s,\*emu,*emu_,' decode.c