Reentrancy: make the external API truly reentrant
Previous commits, while quite invasive, should not have made any
changes to the library's public interfaces, or its externally observable
behaviour.

This commit eliminates the internal global emulator state and makes
the library API actually reentrant. While on the surface it may seem
it doesn't change the API at all, it technically introduces a couple
of incompatibilities:

- The value of the pointer passed to the various user-provided callback
  functions changes.
- `x86emu_run` and `x86emu_clear_log` can no longer be called with a NULL
  pointer as an argument.

While it seems unlikely that code exists in the wild that relies on
these details not changing, the above strictly speaking implies that a
major version bump is necessary.

To ease the transition for users of the library, the latter two functions
have been annotated with __attribute__((nonnull)).

The README.md file has also been updated to remove an obsolete remark.